+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- DscTagging: Added parameter 'BuildNumber'.
+
 ## [0.8.3] - 2023-03-16
 
 ### Changed